backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Google Gemini AI API
google-generativeai==0.3.0

# AI/ML - NLP for ATS
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3

# PDF Generation
reportlab==4.0.7

# Image Processing
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Optional - Database
sqlalchemy==2.0.23

# Optional - Redis
redis==5.0.1